# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.StaffCreate:
    properties:
      branch_id:
        type: string
      name:
        type: string
      tarif_id:
        type: string
      type:
        type: string
    type: object
...
  models.StaffUpdate:
    properties:
      balace:
        type: integer
      branch_id:
        type: string
      id:
        type: string
      name:
        type: string
      tarif_id:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /staff:
    get:
      consumes:
      - application/json
      description: Get List Staff
      operationId: get_list_staff
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      - description: from
        in: query
        name: from
        type: string
      - description: to
        in: query
        name: to
        type: string
      - description: search_by_branch
        in: query
        name: search_by_branch
        type: string
      - description: search_by_tarif
        in: query
        name: search_by_tarif
        type: string
      - description: search_by_type
        in: query
        name: search_by_type
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Staff
      tags:
      - Staff
    post:
      consumes:
      - application/json
      description: Create Staff
      operationId: create_staff
      parameters:
      - description: StaffCreateRequest
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.StaffCreate'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Staff
      tags:
      - Staff
  /staff/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Staff
      operationId: delete_staff
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Staff
      tags:
      - Staff
    get:
      consumes:
      - application/json
      description: Get By ID Staff
      operationId: get_by_id_staff
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Staff
      tags:
      - Staff
    put:
      consumes:
      - application/json
      description: Update Staff
      operationId: update_staff
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: StaffUpdateRequest
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.StaffUpdate'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Staff
      tags:
      - Staff
...
swagger: "2.0"